+++ agent-inject/agent/annotations_test.go
@@ -9,14 +9,20 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 
+	"github.com/hashicorp/vault/sdk/helper/pointerutil"
 	"github.com/mattbaird/jsonpatch"
+	"github.com/stretchr/testify/require"
 )
 
 func TestInitCanSet(t *testing.T) {
 	annotations := make(map[string]string)
 	pod := testPod(annotations)
 
-	err := Init(pod, AgentConfig{"foobar-image", "http://foobar:8200", "test", "test", true, "1000", "100"})
+	agentConfig := AgentConfig{
+		"foobar-image", "http://foobar:8200", "test", "test", true, "100", "1000",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+	}
+	err := Init(pod, agentConfig)
 	if err != nil {
 		t.Errorf("got error, shouldn't have: %s", err)
 	}
@@ -48,7 +54,11 @@ func TestInitDefaults(t *testing.T) {
 	annotations := make(map[string]string)
 	pod := testPod(annotations)
 
-	err := Init(pod, AgentConfig{"", "http://foobar:8200", "test", "test", true, "", ""})
+	agentConfig := AgentConfig{
+		"", "http://foobar:8200", "test", "test", true, "", "",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+	}
+	err := Init(pod, agentConfig)
 	if err != nil {
 		t.Errorf("got error, shouldn't have: %s", err)
 	}
@@ -78,7 +88,11 @@ func TestInitError(t *testing.T) {
 	annotations := make(map[string]string)
 	pod := testPod(annotations)
 
-	err := Init(pod, AgentConfig{"image", "", "authPath", "namespace", true, "1000", "100"})
+	agentConfig := AgentConfig{
+		"image", "", "authPath", "namespace", true, "100", "1000",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+	}
+	err := Init(pod, agentConfig)
 	if err == nil {
 		t.Error("expected error no address, got none")
 	}
@@ -88,7 +102,9 @@ func TestInitError(t *testing.T) {
 		t.Errorf("expected '%s' error, got %s", errMsg, err)
 	}
 
-	err = Init(pod, AgentConfig{"image", "address", "", "namespace", true, "1000", "100"})
+	agentConfig.Address = "address"
+	agentConfig.AuthPath = ""
+	err = Init(pod, agentConfig)
 	if err == nil {
 		t.Error("expected error no authPath, got none")
 	}
@@ -98,7 +114,9 @@ func TestInitError(t *testing.T) {
 		t.Errorf("expected '%s' error, got %s", errMsg, err)
 	}
 
-	err = Init(pod, AgentConfig{"image", "address", "authPath", "", true, "1000", "100"})
+	agentConfig.AuthPath = "authPath"
+	agentConfig.Namespace = ""
+	err = Init(pod, agentConfig)
 	if err == nil {
 		t.Error("expected error for no namespace, got none")
 	}
@@ -134,7 +152,62 @@ func TestSecretAnnotationsWithPreserveCaseSensitivityFlagOff(t *testing.T) {
 		pod := testPod(annotation)
 		var patches []*jsonpatch.JsonPatchOperation
 
-		err := Init(pod, AgentConfig{"", "http://foobar:8200", "test", "test", true, "1000", "100"})
+		agentConfig := AgentConfig{
+			"", "http://foobar:8200", "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		}
+		err := Init(pod, agentConfig)
+		if err != nil {
+			t.Errorf("got error, shouldn't have: %s", err)
+		}
+
+		agent, err := New(pod, patches)
+		if err != nil {
+			t.Errorf("got error, shouldn't have: %s", err)
+		}
+
+		if tt.expectedKey != "" {
+			if len(agent.Secrets) == 0 {
+				t.Error("Secrets length was zero, it shouldn't have been")
+			}
+			if agent.Secrets[0].Name != tt.expectedKey {
+				t.Errorf("expected %s, got %s", tt.expectedKey, agent.Secrets[0].Name)
+			}
+
+			if agent.Secrets[0].Path != tt.expectedPath {
+				t.Errorf("expected %s, got %s", tt.expectedPath, agent.Secrets[0].Path)
+
+			}
+		} else if len(agent.Secrets) > 0 {
+			t.Errorf("Secrets length was greater than zero, it shouldn't have been: %s", tt.key)
+		}
+	}
+}
+
+func TestSecretAnnotationsWithPreserveCaseSensitivityFlagOn(t *testing.T) {
+	tests := []struct {
+		key          string
+		value        string
+		expectedKey  string
+		expectedPath string
+	}{
+		{"vault.hashicorp.com/agent-inject-secret-foobar", "test1", "foobar", "test1"},
+		{"vault.hashicorp.com/agent-inject-secret-FOOBAR", "test2", "FOOBAR", "test2"},
+	}
+
+	for _, tt := range tests {
+		annotation := map[string]string{
+			tt.key:                       tt.value,
+			AnnotationPreserveSecretCase: "true",
+		}
+		pod := testPod(annotation)
+		var patches []*jsonpatch.JsonPatchOperation
+
+		agentConfig := AgentConfig{
+			"", "http://foobar:8200", "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		}
+		err := Init(pod, agentConfig)
 		if err != nil {
 			t.Errorf("got error, shouldn't have: %s", err)
 		}
@@ -162,50 +235,95 @@ func TestSecretAnnotationsWithPreserveCaseSensitivityFlagOff(t *testing.T) {
 	}
 }
 
-func TestSecretAnnotationsWithPreserveCaseSensitivityFlagOn(t *testing.T) {
+func TestSecretLocationFileAnnotations(t *testing.T) {
 	tests := []struct {
-		key          string
-		value        string
-		expectedKey  string
-		expectedPath string
+		name             string
+		annotations      map[string]string
+		expectedName     string
+		expectedFilename string
+		expectedLocation string
 	}{
-		{"vault.hashicorp.com/agent-inject-secret-foobar", "test1", "foobar", "test1"},
-		{"vault.hashicorp.com/agent-inject-secret-FOOBAR", "test2", "FOOBAR", "test2"},
+		{
+			"simple name",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foobar": "vault/test1",
+				"vault.hashicorp.com/agent-inject-file-foobar":   "foobar_simple_name",
+			},
+			"foobar",
+			"foobar_simple_name",
+			"vault/test1",
+		},
+		{
+			"absolute file path",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foobar": "vault/test1",
+				"vault.hashicorp.com/agent-inject-file-foobar":   "/some/path/foobar_simple_name",
+			},
+			"foobar",
+			"/some/path/foobar_simple_name",
+			"vault/test1",
+		},
+		{
+			"long file name",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foobar": "vault/test2",
+				"vault.hashicorp.com/agent-inject-file-foobar":   "this_is_very_long_and/would_fail_in_kubernetes/if_in_annotation",
+			},
+			"foobar",
+			"this_is_very_long_and/would_fail_in_kubernetes/if_in_annotation",
+			"vault/test2",
+		},
+		{
+			"file doesn't match secret annotation",
+			map[string]string{
+				"vault.hashicorp.com/agent-inject-secret-foobar":         "vault/test2",
+				"vault.hashicorp.com/agent-inject-file-notcorresponding": "this_is_very_long_and/would_fail_in_kubernetes/if_in_annotation",
+			},
+			"foobar",
+			"",
+			"vault/test2",
+		},
 	}
 
 	for _, tt := range tests {
-		annotation := map[string]string{
-			tt.key:                       tt.value,
-			AnnotationPreserveSecretCase: "true",
-		}
-		pod := testPod(annotation)
-		var patches []*jsonpatch.JsonPatchOperation
+		t.Run(tt.name, func(t *testing.T) {
+			pod := testPod(tt.annotations)
+			var patches []*jsonpatch.JsonPatchOperation
 
-		err := Init(pod, AgentConfig{"", "http://foobar:8200", "test", "test", true, "1000", "100"})
-		if err != nil {
-			t.Errorf("got error, shouldn't have: %s", err)
-		}
-
-		agent, err := New(pod, patches)
-		if err != nil {
-			t.Errorf("got error, shouldn't have: %s", err)
-		}
-
-		if tt.expectedKey != "" {
-			if len(agent.Secrets) == 0 {
-				t.Error("Secrets length was zero, it shouldn't have been")
+			agentConfig := AgentConfig{
+				"", "http://foobar:8200", "test", "test", true, "100", "1000",
+				DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
 			}
-			if agent.Secrets[0].Name != tt.expectedKey {
-				t.Errorf("expected %s, got %s", tt.expectedKey, agent.Secrets[0].Name)
+			err := Init(pod, agentConfig)
+			if err != nil {
+				t.Errorf("got error, shouldn't have: %s", err)
 			}
 
-			if agent.Secrets[0].Path != tt.expectedPath {
-				t.Errorf("expected %s, got %s", tt.expectedPath, agent.Secrets[0].Path)
-
+			agent, err := New(pod, patches)
+			if err != nil {
+				t.Errorf("got error, shouldn't have: %s", err)
 			}
-		} else if len(agent.Secrets) > 0 {
-			t.Error("Secrets length was greater than zero, it shouldn't have been")
-		}
+
+			if tt.expectedName != "" {
+				if len(agent.Secrets) == 0 {
+					t.Error("Secrets length was zero, it shouldn't have been")
+				}
+
+				if agent.Secrets[0].Name != tt.expectedName {
+					t.Errorf("expected name %s, got %s", tt.expectedName, agent.Secrets[0].Name)
+				}
+
+				if agent.Secrets[0].FilePathAndName != tt.expectedFilename {
+					t.Errorf("expected file %s, got %s", tt.expectedFilename, agent.Secrets[0].Name)
+				}
+
+				if agent.Secrets[0].Path != tt.expectedLocation {
+					t.Errorf("expected path %s, got %s", tt.expectedLocation, agent.Secrets[0].Path)
+				}
+			} else if len(agent.Secrets) > 0 {
+				t.Errorf("Secrets length was greater than zero, it shouldn't have been")
+			}
+		})
 	}
 }
 
@@ -249,7 +367,7 @@ func TestSecretTemplateAnnotations(t *testing.T) {
 			map[string]string{
 				"vault.hashicorp.com/agent-inject-secret-foobar2":  "test1",
 				"vault.hashicorp.com/agent-inject-TEMPLATE-foobar": "foobarTemplate",
-			}, "foobar2", "foobarTemplate",
+			}, "foobar2", "",
 		},
 	}
 
@@ -257,7 +375,11 @@ func TestSecretTemplateAnnotations(t *testing.T) {
 		pod := testPod(tt.annotations)
 		var patches []*jsonpatch.JsonPatchOperation
 
-		err := Init(pod, AgentConfig{"", "http://foobar:8200", "test", "test", true, "1000", "100"})
+		agentConfig := AgentConfig{
+			"", "http://foobar:8200", "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		}
+		err := Init(pod, agentConfig)
 		if err != nil {
 			t.Errorf("got error, shouldn't have: %s", err)
 		}
@@ -275,7 +397,7 @@ func TestSecretTemplateAnnotations(t *testing.T) {
 			t.Errorf("expected name %s, got %s", tt.expectedKey, agent.Secrets[0].Name)
 		}
 
-		if agent.Secrets[0].Name != tt.expectedKey {
+		if agent.Secrets[0].Template != tt.expectedTemplate {
 			t.Errorf("expected template %s, got %s", tt.expectedTemplate, agent.Secrets[0].Template)
 		}
 	}
@@ -313,7 +435,11 @@ func TestTemplateShortcuts(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			pod := testPod(tt.annotations)
-			err := Init(pod, AgentConfig{"", "http://foobar:8200", "test", "test", true, "1000", "100"})
+			agentConfig := AgentConfig{
+				"", "http://foobar:8200", "test", "test", true, "100", "1000",
+				DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+			}
+			err := Init(pod, agentConfig)
 			if err != nil {
 				t.Errorf("got error, shouldn't have: %s", err)
 			}
@@ -369,7 +495,11 @@ func TestSecretCommandAnnotations(t *testing.T) {
 
 	for _, tt := range tests {
 		pod := testPod(tt.annotations)
-		err := Init(pod, AgentConfig{"", "http://foobar:8200", "test", "test", true, "1000", "100"})
+		agentConfig := AgentConfig{
+			"", "http://foobar:8200", "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		}
+		err := Init(pod, agentConfig)
 		if err != nil {
 			t.Errorf("got error, shouldn't have: %s", err)
 		}
@@ -466,6 +596,7 @@ func TestCouldErrorAnnotations(t *testing.T) {
 		{AnnotationAgentRevokeGrace, "01", true},
 		{AnnotationAgentRevokeGrace, "-1", false},
 		{AnnotationAgentRevokeGrace, "foobar", false},
+
 		{AnnotationAgentRunAsUser, "0", true},
 		{AnnotationAgentRunAsUser, "100", true},
 		{AnnotationAgentRunAsUser, "root", false},
@@ -473,6 +604,23 @@ func TestCouldErrorAnnotations(t *testing.T) {
 		{AnnotationAgentRunAsGroup, "0", true},
 		{AnnotationAgentRunAsGroup, "100", true},
 		{AnnotationAgentRunAsGroup, "root", false},
+
+		{AnnotationAgentSetSecurityContext, "true", true},
+		{AnnotationAgentSetSecurityContext, "false", true},
+		{AnnotationAgentSetSecurityContext, "secure", false},
+		{AnnotationAgentSetSecurityContext, "", false},
+
+		{AnnotationAgentCacheEnable, "true", true},
+		{AnnotationAgentCacheEnable, "false", true},
+		{AnnotationAgentCacheEnable, "TRUE", true},
+		{AnnotationAgentCacheEnable, "FALSE", true},
+		{AnnotationAgentCacheEnable, "0", true},
+		{AnnotationAgentCacheEnable, "1", true},
+		{AnnotationAgentCacheEnable, "t", true},
+		{AnnotationAgentCacheEnable, "f", true},
+		{AnnotationAgentCacheEnable, "tRuE", false},
+		{AnnotationAgentCacheEnable, "fAlSe", false},
+		{AnnotationAgentCacheEnable, "", false},
 	}
 
 	for i, tt := range tests {
@@ -480,7 +628,11 @@ func TestCouldErrorAnnotations(t *testing.T) {
 		pod := testPod(annotations)
 		var patches []*jsonpatch.JsonPatchOperation
 
-		err := Init(pod, AgentConfig{"", "http://foobar:8200", "test", "test", true, "1000", "100"})
+		agentConfig := AgentConfig{
+			"", "http://foobar:8200", "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		}
+		err := Init(pod, agentConfig)
 		if err != nil {
 			t.Errorf("got error, shouldn't have: %s", err)
 		}
@@ -497,7 +649,11 @@ func TestCouldErrorAnnotations(t *testing.T) {
 func TestInitEmptyPod(t *testing.T) {
 	var pod *corev1.Pod
 
-	err := Init(pod, AgentConfig{"foobar-image", "http://foobar:8200", "test", "test", true, "1000", "100"})
+	agentConfig := AgentConfig{
+		"foobar-image", "http://foobar:8200", "test", "test", true, "100", "1000",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+	}
+	err := Init(pod, agentConfig)
 	if err == nil {
 		t.Errorf("got no error, should have")
 	}
@@ -522,7 +678,11 @@ func TestVaultNamespaceAnnotation(t *testing.T) {
 		pod := testPod(annotation)
 		var patches []*jsonpatch.JsonPatchOperation
 
-		err := Init(pod, AgentConfig{"foobar-image", "http://foobar:8200", "test", "test", true, "1000", "100"})
+		agentConfig := AgentConfig{
+			"foobar-image", "http://foobar:8200", "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		}
+		err := Init(pod, agentConfig)
 		if err != nil {
 			t.Errorf("got error, shouldn't have: %s", err)
 		}
@@ -535,5 +695,101 @@ func TestVaultNamespaceAnnotation(t *testing.T) {
 		if agent.Vault.Namespace != tt.expectedValue {
 			t.Errorf("expected %s, got %s", tt.expectedValue, agent.Vault.Namespace)
 		}
+	}
+}
+
+func Test_runAsSameID(t *testing.T) {
+
+	tests := []struct {
+		name           string
+		runAsSameUser  string
+		appSCC         *corev1.SecurityContext
+		expectedResult bool
+		expectedErr    bool
+		expectedUserID int64
+	}{
+		{
+			name:           "false with no app SCC",
+			runAsSameUser:  "false",
+			appSCC:         nil,
+			expectedResult: false,
+			expectedErr:    false,
+			expectedUserID: DefaultAgentRunAsUser,
+		},
+		{
+			name:          "true with app SCC",
+			runAsSameUser: "true",
+			appSCC: &corev1.SecurityContext{
+				RunAsUser: pointerutil.Int64Ptr(123456),
+			},
+			expectedResult: true,
+			expectedErr:    false,
+			expectedUserID: 123456,
+		},
+		{
+			name:          "false with app SCC",
+			runAsSameUser: "false",
+			appSCC: &corev1.SecurityContext{
+				RunAsUser: pointerutil.Int64Ptr(123456),
+			},
+			expectedResult: false,
+			expectedErr:    false,
+			expectedUserID: DefaultAgentRunAsUser,
+		},
+		{
+			name:           "true with no app SCC",
+			runAsSameUser:  "true",
+			appSCC:         nil,
+			expectedResult: false,
+			expectedErr:    true,
+			expectedUserID: DefaultAgentRunAsUser,
+		},
+		{
+			name:           "annotation not set",
+			runAsSameUser:  "",
+			appSCC:         nil,
+			expectedResult: false,
+			expectedErr:    false,
+			expectedUserID: DefaultAgentRunAsUser,
+		},
+		{
+			name:           "invalid annotation set",
+			runAsSameUser:  "rooooooot",
+			appSCC:         nil,
+			expectedResult: false,
+			expectedErr:    true,
+			expectedUserID: DefaultAgentRunAsUser,
+		},
+		{
+			name:          "true with app SCC as root user",
+			runAsSameUser: "true",
+			appSCC: &corev1.SecurityContext{
+				RunAsUser: pointerutil.Int64Ptr(0),
+			},
+			expectedResult: false,
+			expectedErr:    true,
+			expectedUserID: DefaultAgentRunAsUser,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+
+			annotations := map[string]string{}
+			if len(tt.runAsSameUser) > 0 {
+				annotations[AnnotationAgentRunAsSameUser] = tt.runAsSameUser
+			}
+			pod := testPod(annotations)
+			pod.Spec.Containers[0].SecurityContext = tt.appSCC
+
+			agent := &Agent{
+				Annotations: annotations,
+				RunAsUser:   DefaultAgentRunAsUser,
+			}
+			result, err := agent.runAsSameID(pod)
+			require.Equal(t, tt.expectedResult, result)
+			require.Equal(t, tt.expectedErr, err != nil)
+			require.Equal(t, tt.expectedUserID, agent.RunAsUser)
+		})
 	}
 }