+++ agent-inject/agent/container_init_sidecar.go
@@ -14,7 +14,7 @@ import (
 func (a *Agent) ContainerInitSidecar() (corev1.Container, error) {
 	volumeMounts := []corev1.VolumeMount{
 		{
-			Name:      tokenVolumeName,
+			Name:      tokenVolumeNameInit,
 			MountPath: tokenVolumePath,
 			ReadOnly:  false,
 		},
@@ -25,6 +25,14 @@ func (a *Agent) ContainerInitSidecar() (corev1.Container, error) {
 		},
 	}
 	volumeMounts = append(volumeMounts, a.ContainerVolumeMounts()...)
+
+	if a.ExtraSecret != "" {
+		volumeMounts = append(volumeMounts, corev1.VolumeMount{
+			Name:      extraSecretVolumeName,
+			MountPath: extraSecretVolumePath,
+			ReadOnly:  true,
+		})
+	}
 
 	arg := DefaultContainerArg
 
@@ -55,14 +63,18 @@ func (a *Agent) ContainerInitSidecar() (corev1.Container, error) {
 		return corev1.Container{}, err
 	}
 
-	return corev1.Container{
-		Name:            "vault-agent-init",
-		Image:           a.ImageName,
-		Env:             envs,
-		Resources:       resources,
-		SecurityContext: a.securityContext(),
-		VolumeMounts:    volumeMounts,
-		Command:         []string{"/bin/sh", "-ec"},
-		Args:            []string{arg},
-	}, nil
+	newContainer := corev1.Container{
+		Name:         "vault-agent-init",
+		Image:        a.ImageName,
+		Env:          envs,
+		Resources:    resources,
+		VolumeMounts: volumeMounts,
+		Command:      []string{"/bin/sh", "-ec"},
+		Args:         []string{arg},
+	}
+	if a.SetSecurityContext {
+		newContainer.SecurityContext = a.securityContext()
+	}
+
+	return newContainer, nil
 }