+++ agent-inject/agent/container_volume.go
@@ -8,14 +8,17 @@ import (
 )
 
 const (
-	tokenVolumeName     = "home"
-	tokenVolumePath     = "/home/vault"
-	configVolumeName    = "vault-config"
-	configVolumePath    = "/vault/configs"
-	secretVolumeName    = "vault-secrets"
-	tlsSecretVolumeName = "vault-tls-secrets"
-	tlsSecretVolumePath = "/vault/tls"
-	secretVolumePath    = "/vault/secrets"
+	tokenVolumeNameInit    = "home-init"
+	tokenVolumeNameSidecar = "home-sidecar"
+	tokenVolumePath        = "/home/vault"
+	configVolumeName       = "vault-config"
+	configVolumePath       = "/vault/configs"
+	secretVolumeName       = "vault-secrets"
+	tlsSecretVolumeName    = "vault-tls-secrets"
+	tlsSecretVolumePath    = "/vault/tls"
+	secretVolumePath       = "/vault/secrets"
+	extraSecretVolumeName  = "extra-secrets"
+	extraSecretVolumePath  = "/vault/custom"
 )
 
 func (a *Agent) getUniqueMountPaths() []string {
@@ -60,15 +63,32 @@ func (a *Agent) ContainerVolumes() []corev1.Volume {
 
 // ContainerTokenVolume returns a volume to mount the
 // home directory where the token sink will write to.
-func (a *Agent) ContainerTokenVolume() corev1.Volume {
-	return corev1.Volume{
-		Name: tokenVolumeName,
-		VolumeSource: corev1.VolumeSource{
-			EmptyDir: &corev1.EmptyDirVolumeSource{
-				Medium: "Memory",
+func (a *Agent) ContainerTokenVolume() []corev1.Volume {
+	var vols []corev1.Volume
+	if a.PrePopulate {
+		initVol := corev1.Volume{
+			Name: tokenVolumeNameInit,
+			VolumeSource: corev1.VolumeSource{
+				EmptyDir: &corev1.EmptyDirVolumeSource{
+					Medium: "Memory",
+				},
 			},
-		},
+		}
+		vols = append(vols, initVol)
 	}
+	if !a.PrePopulateOnly {
+		sidecarVol := corev1.Volume{
+			Name: tokenVolumeNameSidecar,
+			VolumeSource: corev1.VolumeSource{
+				EmptyDir: &corev1.EmptyDirVolumeSource{
+					Medium: "Memory",
+				},
+			},
+		}
+		vols = append(vols, sidecarVol)
+	}
+
+	return vols
 }
 
 // ContainerConfigMapVolume returns a volume to mount a config map
@@ -81,6 +101,19 @@ func (a *Agent) ContainerConfigMapVolume() corev1.Volume {
 				LocalObjectReference: corev1.LocalObjectReference{
 					Name: a.ConfigMapName,
 				},
+			},
+		},
+	}
+}
+
+// ContainerExtraSecretVolume returns a volume to mount a Kube secret
+// if the user supplied one.
+func (a *Agent) ContainerExtraSecretVolume() corev1.Volume {
+	return corev1.Volume{
+		Name: extraSecretVolumeName,
+		VolumeSource: corev1.VolumeSource{
+			Secret: &corev1.SecretVolumeSource{
+				SecretName: a.ExtraSecret,
 			},
 		},
 	}